requirements.txt
annotated-types==0.7.0
anyio==4.9.0
audioread==3.0.1
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
decorator==5.1.1
dnspython==2.7.0
fastapi==0.115.12
h11==0.14.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
joblib==1.4.2
lazy_loader==0.4
llvmlite==0.44.0
msgpack==1.1.0
numba==0.61.0
numpy==2.1.3
packaging==24.2
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pymongo==4.11
PyMuPDF==1.25.2
python-dotenv==1.0.1
python-multipart==0.0.20
RapidFuzz==3.12.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.1
sniffio==1.3.1
soxr==0.5.0.post1
srt==3.5.3
starlette==0.46.1
threadpoolctl==3.5.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
vosk==0.3.44
websockets==14.2
